requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
rapidfuzz==3.9.7
python-dotenv==1.0.1
gunicorn==22.0.0
